Add utility script for PyEnv PRs
This image uses PyEnv to install and manage Python versions. Sometimes
PyEnv doesn't get Python versions fast enough and someone from CPE will
submit the PR to the project. In that process, a new file needs to be
created with sha256 hashes for the Python tarball download.

Their docs doesn't tell you that it's sha256 and it has to be done
manually. This PR adds a script to our repo tho spit out the hashes for
you so that you can just copy and paste them. You'd use it by running:

python-hashes.sh <python-version>

for example:

`python-hashes.sh 3.7.10`